Create the root reducer
